12-Web-Scraping-and-Document-Databases/webscraping/requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.4.1
numpy==1.16.4
pandas==0.24.2
progressbar2==3.42.0
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
pymongo==3.9.0
progressbar2==3.42.0 
